python/requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=10.0.0
